+++ test/spec/default.spec.ts
@@ -1,0 +1,9 @@
+import { Log } from '../../src';
+import { log } from '../../src/default';
+
+describe('default', () => {
+	it('exports a default log instance', () => {
+		expect(log).toBeDefined();
+		expect(log instanceof Log).toBeTrue();
+	});
+});