+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ts-tiny-log
 
+## [1.0.2]
+
+### Fixes
+- Export default instance from `/default`
+
 ## [1.0.1]
 
 Initial Release